firstFaissTestProgram/requirements.txt
faiss-cpu
torch 
torchvision
ftfy 
regex 
tqdm
git+https://githubfast.com/openai/CLIP.git
matplotlib
numpy
pillow